+++ skyportal/tests/external/test_catalog_services.py
